Some struct tag clean up and prepared statement testing
Struct tags are pretty much what they should be now. Go is PascalCase so that it can exported, json is camelCase, and db is snake_case. I came to this conclusion after looking into what is to be considered as "industry standard" for each "language".

Need to test prepared statements further.