This commit adds cursory support for web session persistence.
We've added a new token to be granted (which pins us to only one instance
of the backend to run at a time--this will be fixed when we add support
for redis as a central store and create hard tokens. This token allows
for end users to communicate with the backend to manage their apps.